Parse multiple files without headers
Previously it was not possible to parse a multi-file diff unless they
were separated by headers (e.g., `diff --git`). Since headers are
optional and not produced by a simple `diff -u`, this prevents some
valid diffs from being parsed.

Additionally, when reading a multi-file diff line-by-line, it is
difficult to determine if a line starting with `---` is (1) the start of
a new file (2) removal of a line starting with `--` (two hyphens). The
solution in this commit uses the line counts embedded in the hunk header
to determine when the hunk is finished. That way, we don't have to
resort to messier solutions such as reading ahead by multiple lines to
try to make the determination. The test data `sample_bad_hunks.diff`
previously asserted that this library could parse a diff with incorrect
line counts in the hunk header. With this change, parsing that file or a
similar one with incorrect line counts will fail.