Fix parsing of non-extended file headers and lines starting with --
This fixes #54 by making the detection of non-extended file headers
(which start with `---` directly) that was introduced in #53 more
robust.

Instead of simply aborting when the current line starts with `---`
(which is a valid hunk line, if you, say, remove a line starting with
`--`), we confirm that the next line also starts with `+++` by peeking a
bit ahead.

That's also what `git` does: https://sourcegraph.com/github.com/git/git/-/blob/apply.c#L1574-1576