Add support for complicated filenames and file copies
When filenames contain special characters, they are quoted, but not if
they only contain spaces, meaning the 'git --diff' argument line becomes
AMBIGUOUS. It is, however, possible to reconstruct the filenames using
the extended headers.

This fixes parsing of unquoted filenames with multiple spaces in them,
quoted filenames with spaces in them, and diffs containing file copies.